clean: moved intefaces to types.ts
