+++ src/@types/types.ts
@@ -5,3 +5,24 @@ export interface IError {
 export interface IInput {
   type: string;
 }
+
+
+export interface IUser {
+  email: string;
+  name: string;
+}
+
+export interface IPost {
+  imgUrl: string;
+  created_by: string;
+  created_at: string;
+  likes: string[];
+  sent: string[];
+  waittingAnswer: string[];
+  comments: Comment[];
+}
+
+export interface Comment {
+  user: string;
+  comment: string;
+}
